// contracts/types/protocol/gig/Gig.ts
export declare namespace DataTypes {
  export type PackageStruct = {
    price: PromiseOrValue<BigNumberish>;
    timeDelivery: PromiseOrValue<BigNumberish>;
  };

  export type PackageStructOutput = [BigNumber, BigNumber] & {
    price: BigNumber;
    timeDelivery: BigNumber;
  };
}

export declare namespace OutputTypes {
  export type GigOutputStruct = {
    gigId: PromiseOrValue<BigNumberish>;
    metadata: PromiseOrValue<string>;
    createdAt: PromiseOrValue<BigNumberish>;
    successSell: PromiseOrValue<BigNumberish>;
    failedSell: PromiseOrValue<BigNumberish>;
    reviewIds: PromiseOrValue<BigNumberish>[];
    orderIds: PromiseOrValue<BigNumberish>[];
    packages: [
      DataTypes.PackageStruct,
      DataTypes.PackageStruct,
      DataTypes.PackageStruct
    ];
  };

  export type GigOutputStructOutput = [
    BigNumber,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber[],
    BigNumber[],
    [
      DataTypes.PackageStructOutput,
      DataTypes.PackageStructOutput,
      DataTypes.PackageStructOutput
    ]
  ] & {
    gigId: BigNumber;
    metadata: string;
    createdAt: BigNumber;
    successSell: BigNumber;
    failedSell: BigNumber;
    reviewIds: BigNumber[];
    orderIds: BigNumber[];
    packages: [
      DataTypes.PackageStructOutput,
      DataTypes.PackageStructOutput,
      DataTypes.PackageStructOutput
    ];
  };
}